test(integration): address PR #2491 review feedback
- package.json: fix test:streaming script (.mts -> .ts)
- background/api/multitasking/streaming tests: drop retry: 3 (keep timeouts)
- streaming.test.ts: move readSSEStream helper into utils.ts
- README.md: replace PR-specific rationale with general docs
- compose-postgres.yml: drop unused FF_JS_ZEROMQ_ENABLED passthrough
- agent_simple.ts: use StateSchema + MessagesValue instead of Annotation.Root
