+++ integration/package.json
@@ -9,7 +9,7 @@
     "up": "docker compose -f compose-postgres.yml up -V --force-recreate --wait",
     "down": "docker compose -f compose-postgres.yml down -v",
     "test": "vitest run --fileParallelism=false",
-    "test:streaming": "vitest run --fileParallelism=false streaming.test.mts",
+    "test:streaming": "vitest run --fileParallelism=false streaming.test.ts",
     "int": "pnpm configure && pnpm build && pnpm up && (pnpm test; ec=$?; pnpm down; exit $ec)"
   },
   "devDependencies": {
