+++ integration/streaming.test.ts
@@ -1,7 +1,7 @@
 import { Client } from "@langchain/langgraph-sdk";
 import { beforeAll, describe, expect, it } from "vitest";
 import postgres from "postgres";
-import { gatherIterator, findLast } from "./utils.ts";
+import { gatherIterator, findLast, readSSEStream } from "./utils.ts";
 
 const sql = postgres(
   process.env.POSTGRES_URI ??
@@ -26,89 +26,8 @@ beforeAll(async () => {
   await sql`DELETE FROM assistant WHERE metadata->>'created_by' is null OR metadata->>'created_by' != 'system'`;
 });
 
-/**
- * Helper to parse SSE chunks from a raw ReadableStream.
- * Returns an array of { event, data } objects.
- */
-async function readSSEStream(
-  reader: ReadableStreamDefaultReader<Uint8Array>,
-  {
-    timeout = 15_000,
-    stopWhen,
-  }: {
-    timeout?: number;
-    stopWhen?: (events: Array<{ event: string; data: any }>) => boolean;
-  } = {},
-): Promise<Array<{ event: string; data: any }>> {
-  const decoder = new TextDecoder();
-  const events: Array<{ event: string; data: any }> = [];
-  let buffer = "";
-  let currentEvent = "";
-  let currentData = "";
-
-  const deadline = Date.now() + timeout;
-
-  function pushEvent() {
-    if (currentEvent || currentData) {
-      let parsed: any = currentData;
-      try {
-        parsed = JSON.parse(currentData);
-      } catch {
-        // keep as string
-      }
-      events.push({ event: currentEvent, data: parsed });
-      currentEvent = "";
-      currentData = "";
-    }
-  }
-
-  while (Date.now() < deadline) {
-    let timer: ReturnType<typeof setTimeout> | undefined;
-    const readPromise = reader.read();
-    const timeoutPromise = new Promise<{ done: true; value: undefined }>(
-      (resolve) => {
-        timer = setTimeout(
-          () => resolve({ done: true, value: undefined }),
-          Math.max(deadline - Date.now(), 0),
-        );
-      },
-    );
-
-    const { done, value } = await Promise.race([readPromise, timeoutPromise]);
-    clearTimeout(timer);
-    if (done) break;
-
-    buffer += decoder.decode(value, { stream: true });
-
-    // Normalize \r\n → \n (server sends \r\n per SSE spec)
-    buffer = buffer.replace(/\r\n/g, "\n").replace(/\r/g, "\n");
-
-    // Parse SSE frames from buffer
-    const lines = buffer.split("\n");
-    buffer = lines.pop() ?? "";
-
-    for (const line of lines) {
-      if (line.startsWith("event:")) {
-        currentEvent = line.slice(6).trim();
-      } else if (line.startsWith("data:")) {
-        currentData = line.slice(5).trim();
-      } else if (line === "") {
-        pushEvent();
-      }
-    }
-
-    if (stopWhen && stopWhen(events)) break;
-  }
-
-  // Flush any partially accumulated SSE message
-  pushEvent();
-
-  reader.cancel().catch(() => {});
-  return events;
-}
-
 describe("streaming edge cases", () => {
-  it.concurrent("debug stream mode", { retry: 3 }, async () => {
+  it.concurrent("debug stream mode", async () => {
     const assistant = await client.assistants.create({ graphId: "agent" });
     const thread = await client.threads.create();
     const input = {
@@ -148,7 +67,6 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "debug + checkpoints combined",
-    { retry: 3 },
     async () => {
       const assistant = await client.assistants.create({ graphId: "agent" });
       const thread = await client.threads.create();
@@ -195,7 +113,6 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "messages metadata schema",
-    { retry: 3 },
     async () => {
       const assistant = await client.assistants.create({ graphId: "agent" });
       const thread = await client.threads.create();
@@ -238,7 +155,6 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "stream events with subgraphs",
-    { retry: 3 },
     async () => {
       const assistant = await client.assistants.create({
         graphId: "weather",
@@ -288,7 +204,7 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "resumable streams with subgraphs",
-    { retry: 3, timeout: 15_000 },
+    { timeout: 15_000 },
     async () => {
       const assistant = await client.assistants.create({
         graphId: "weather",
@@ -366,7 +282,7 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "thread stream basic",
-    { retry: 3, timeout: 20_000 },
+    { timeout: 20_000 },
     async () => {
       const thread = await client.threads.create();
 
@@ -440,7 +356,7 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "thread stream lifecycle mode",
-    { retry: 3, timeout: 20_000 },
+    { timeout: 20_000 },
     async () => {
       const thread = await client.threads.create();
 
@@ -541,7 +457,6 @@ describe("streaming edge cases", () => {
 
   it.concurrent(
     "stream echo preserves newlines",
-    { retry: 3 },
     async () => {
       const assistant = await client.assistants.create({
         graphId: "agent_simple",
